Updated json condenser to handle nested array, added copy to clipboard button, and updated related tests
